requirements.txt
streamlit
streamlit-option-menu
streamlit-extras
python-dotenv
PyPDF2
langchain
langchain-google-genai
google-generativeai
faiss-cpu
selenium
pandas
numpy
python-docx